Introduced support for pushing raw data.
Some files can be huge and the json encoder can be really
slow on tables with many data so in this cases we need
to generate the json string to send manually.

The change allows sending the textDocument/didOpen on large
files without stalling the whole UI waiting for the json encoder
to finish. It also creates a thread to send the big files
in order to also not block the UI.

This work was tested on sqlite3.c amalgamated source file which
is around 8MB in size and the clangd lsp server.