Raw document save and change notifications.
When a save or change notification needs to send the whole
document content we now do it as raw generated json for
performance reasons with huge files.

Also played a bit with emmylua which is a style of code
annotation supported by the lua language server.

Converted the server object to a Lite Object.