Added support to overwrite not processed requests.
This change was applied to requests, notifications and raw requests. The
change prevents sending too many requests of the same kind like those
emitted when changing the document or asking for signature and
autocompletion.

The amount of requests sent to the language server is significantly
reduced improving the performance, also fixes issues when sending
document updates/changes since they are now consolidated when a document
notification hasn't been already sent.