kludge to fix multiple overloads
Currently, when I'm using jdtls, if I have multiple overloads of
the same method, only one will show up in the list. This is because the
different overloads will all have the same label. This commit adds
invisible characters to the ends of the labels, to make them unique.
That's a hideous solution, but it does work.